pyproject.toml
[tool.poetry]
name = "collect-shapes"
version = "0.1.0"
description = ""
authors = ["PC2 <ahriman2@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pygame-ce = "^2.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
